refactor: move active-list operations into their callers
settings.ts keeps only storage primitives (get/update/subscribe/reconcile).
addToActiveList, editActiveList, removeFromActiveList, and
restoreActiveListDefaults move into OptionsController as private methods;
popup.ts inlines its single blacklist-add write directly.
